refactor: Make Web3 modals more reusable.
