+++ src/components/Guilds/Web3Modals/components/LiveIndicator.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const LiveIndicator = styled.span`
+  height: 8px;
+  width: 8px;
+  margin-right: 0.5rem;
+  background-color: #4cc7a2;
+  border-radius: 50%;
+`;
+
+export default LiveIndicator;
