+++ src/components/Guilds/common/Modal/index.tsx
@@ -38,7 +38,7 @@ export const StyledModal = styled.div`
   position: relative;
   margin: auto;
   border: 1px solid ${({ theme }) => theme.colors.muted};
-  border-radius: ${({ theme }) => theme.radii.curved};
+  border-radius: ${({ theme }) => theme.radii.curved2};
   box-shadow: 0px 0px 8px rgba(0, 0, 0, 0.25);
 `;
 export const Header = styled.div`
